chore: get tvl and volume for key metric table
