+++ apps/hubble-stats/data/reusable/index.ts
@@ -1,0 +1,2 @@
+export { default as getTvl } from './getTvl';
+export { default as getVolume24h } from './getVolume24h';
